pyproject.toml
[project]
name = "discord-command-runner"
version = "0.1.0"
description = "Discord bot that spawns interactive Claude Code sessions via slash commands"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4,<3",
    "pyyaml>=6",
]

[project.scripts]
command-runner = "command_runner.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts/command_runner"]
